refactor: Replaced user-based auth context for actor in bearer and session hooks
